requirements.txt
# 依赖项

# HTTP 客户端
requests>=2.31.0

# 配置和工具
pyyaml>=6.0

# 开发依赖
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0
